requirements.txt
biopython>=1.79
pandas>=1.3.0
numpy>=1.20.0
primer3-py>=0.6.1
tqdm>=4.60.0
python-Levenshtein>=0.12.2